Change the order of locks.
This addresses part of #310.

Make MySQL clean up its locks by rolling back the transaction in the stored proc. This fixes #313.
